server/requirements.txt
bcrypt>=4.0.0
uvicorn>=0.20.0
fastapi>=0.100.0
apscheduler>=3.10.0
httpx>=0.24.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
psycopg[binary]>=3.0.0
pydantic>=2.0.0
